requirements.txt
cachetools==5.5.2
certifi==2025.4.26
chardet==3.0.4
charset-normalizer==3.4.2
google-api-core==2.25.0
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-translate==3.20.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.1
grpcio-status==1.72.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==2.2.6
pandas==2.2.3
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rfc3986==1.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tzdata==2025.2
urllib3==2.4.0
